requirements.txt
pandas 
numpy 
sqlalchemy 
pymysql
mysql-connector-python
